Adding insertion sort algorithm and style changes.
Following Cormen book, I implemented the instertion sort algorithm.
For the following algorithms, I need to pay more attention to the 1
convention used in the book ! My dear, I miss FORTRAN so much ....

The rest were all style changes.

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
    new file:   in_sort.c
    modified:   neat.c
    modified:   v_mult.c